Payment in progress...

+++ src/app/database/migrations/20251114070214_create_ad_campaigns_table.js
@@ -4,27 +4,27 @@ module.exports = {
     up: async (queryInterface, Sequelize) => {
         await queryInterface.sequelize.query(`
             CREATE TABLE IF NOT EXISTS ad_campaigns (
-                id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
-                user_id BIGINT,
-                published BOOLEAN DEFAULT FALSE,
-                is_matching BOOLEAN DEFAULT FALSE,
-                is_test BOOLEAN DEFAULT FALSE,
-                name VARCHAR(255) NOT NULL,
-                platform VARCHAR(32),
-                follower_min INTEGER,
-                follower_max INTEGER,
-                likes_min INTEGER,
-                likes_max INTEGER,
-                content_link VARCHAR NULL,
+                id              BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+                user_id         BIGINT,
+                published       BOOLEAN DEFAULT FALSE,
+                is_matching     BOOLEAN DEFAULT FALSE,
+                is_test         BOOLEAN DEFAULT FALSE,
+                name            VARCHAR(255) NOT NULL,
+                platform        VARCHAR(32),
+                follower_min    INTEGER,
+                follower_max    INTEGER,
+                likes_min       INTEGER,
+                likes_max       INTEGER,
+                content_link    VARCHAR NULL,
                 ad_campaign_engagement_range_id INTEGER NOT NULL,
-                publish_from timestamptz NOT NULL DEFAULT NOW(),
-                publish_until timestamptz,
+                publish_from    timestamptz NOT NULL DEFAULT NOW(),
+                publish_until   timestamptz,
                 ad_campaign_deliverable_id INTEGER NOT NULL,
-                genre_id INTEGER,
+                genre_id        INTEGER,
                 ad_campaign_payment_type_id INTEGER,
-                price DECIMAL(10,2) DEFAULT 0.00 CHECK (price >= 0),
-                description TEXT,
-                archived BOOLEAN DEFAULT FALSE,
+                price integer   DEFAULT 0.00 CHECK (price >= 0),
+                description     TEXT,
+                archived        BOOLEAN DEFAULT FALSE,
                 impressions_cap BIGINT DEFAULT NULL,
                 story_impressions_min INTEGER DEFAULT NULL,
                 story_impressions_max INTEGER DEFAULT NULL,
@@ -32,11 +32,9 @@ module.exports = {
                 ad_campaign_genre_id BIGINT,
                 meta JSON,
                 impressions_cap_state INTEGER DEFAULT 0,
-                
-                draft_date timestamptz,
-                
-                created_at timestamptz,
-                updated_at timestamptz
+                draft_date      timestamptz,
+                created_at      timestamptz,
+                updated_at      timestamptz
             );
 
             -- Indexes for performance
